Move terminal configuration into new common file
